larger datatypes for the quotients, and trying to compute intermediary steps to no avail

+++ include/my_linalg.h
@@ -18,7 +18,7 @@ quota quota_sub(quota a, quota b);
 
 float quota_fcompute(quota q);
 
-int quota_icompute(quota q);
+int_fast64_t quota_icompute(quota q);
 
 bool quota_is_int(quota q);
 #endif
